refactor: 이벤트 네이밍 정확히 명시
