+++ src/main/java/com/gloddy/server/mate/event/MateCreateEvent.java
@@ -1,4 +1,4 @@
-package com.gloddy.server.praise.event;
+package com.gloddy.server.mate.event;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PraiseCountUpdateEvent {
-    private Long userId;
-    private boolean isAbsenceCountUpdate;
+public class MateCreateEvent {
+    private Long matedUserId;
 }
